Get rid of the package
