+++ lambdas/transcode/setup.py
@@ -3,6 +3,5 @@ from setuptools import setup
 setup(
     name='t4_lambda_transcode',
     version='0.0.1',
-    packages=['transcode'],
     py_modules=['index'],
 )
